exercise-13: add logging utility
