+++ todo_app/utility/logging_utility.py
@@ -1,0 +1,7 @@
+import logging
+
+def configure_logs():
+    # Apply global logging configuration.
+    # Run this at the root of every function.
+
+    logging.getLogger("azure_todo_app").setLevel(logging.WARNING)
